Add coverage of privileged unit tests (#1808)
* Add coverage of privileged tests

* rename new GH action

* Address copilot code review

* test-privileged runs ONLY privileged tests

* recompiled tc_program

* fix race condition

* Fix another race condition in tcmanager_test

* Revert most of common_test.go

* fix unit testsgit push mariomac main:add-priv-cov

* fix BPF verifier tesst

* fix permission in bpf-verifier test

* Really fixing bpf verifier tests :)

* bpf tests create test output folder

* BPF verifier tests don't really need reporting coverage data

* Fix linter
