+++ .github/workflows/pull_request.yml
@@ -282,4 +282,4 @@ jobs:
 
       - name: Run BPF verifier tests (host kernel)
         if: ${{ !matrix.use_vm }}
-        run: sudo env PATH=$PATH PRIVILEGED_TESTS=true go test -v -count=1 ./pkg/internal/ebpf/verifier/...
+        run: go test -exec=sudo -v -count=1 -tags=bpf_verifier_tests ./pkg/internal/ebpf/verifier/...
